requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
